Fix : id return type Integer automatically cast to long type in jjwt

+++ ollive_back/src/main/java/org/palpalmans/ollive_back/domain/member/service/JwtService.java
@@ -58,8 +58,7 @@ public class JwtService {
     }
 
     public long getMemberId(String token){
-        //todo : Check if the code is well-written
-       return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("id", long.class);
+       return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("id", Integer.class);
     }
 
 }
